Fixing test for s3 bucket target class

+++ test/s3_data_packer/targets/s3_bucket_test.rb
@@ -24,7 +24,7 @@ class S3BucketTargetTest < Minitest::Test
       @target.save_file file.path
       assert @fake_resource.bucket('test').store.keys.size == 1
       assert @fake_resource.bucket('test').store.values.first == 'asd qwe'
-      File.delete(file.path)
+      File.delete(file.path) if File.exist?(file.path)
     end
   end
